+++ src/main/java/br/lucas/bp/drogaria/repository/CategoriaRepository.java
@@ -6,6 +6,8 @@ import br.lucas.bp.drogaria.domain.Categoria;
 
 /*Id da categoria é Byte
  * CategoriaRepository  tem editar, salvar, excluir, listar, buscar ...*/
-public interface CategoriaRepository extends JpaRepository<Categoria, Byte>{
+/*Mudei Byte para Short, a geração da chave primaria auto incremento no está implementada para Byte
+ * só implementada a partir do short motivo JPA*/
+public interface CategoriaRepository extends JpaRepository<Categoria, Short>{
 
 }
