+++ src/main/java/br/lucas/bp/drogaria/repository/ProdutoRepository.java
@@ -8,6 +8,6 @@ import br.lucas.bp.drogaria.domain.Produto;
  * JpaRepository: Ele remete a persistencia generica que costuma codificar antigamente
  * Ele tem o buscar, buscarAll,findAllById: encontrar todos pela chave primaria, salvar......
  * Short: chave primaria do produto.Class*/
-public interface ProdutoRepository extends JpaRepository<Produto, Short>{
+public interface ProdutoRepository extends JpaRepository<Produto, Integer>{
 
 }
